Server: properly kill players.
Pipes use POLLHUP when the write end is closed, unlike sockets, which
just use POLLIN; change the code to check for POLLHUP events.  Also,
finish reading data from the server and send it out to the player (if
possible) before closing the connection; previously was trying to read
data in from the player, which should only happen when the player hangs
up, not when they are killed.